+++ src/AddressToken.php
@@ -1,0 +1,37 @@
+<?php
+
+declare(strict_types=1);
+/*
+ * UrlLinker converts any web addresses in plain text into HTML hyperlinks.
+ * Copyright (C) 2016-2025  Artur Weigandt  <https://wlabs.de/kontakt>
+
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+namespace Art4\UrlLinker;
+
+/**
+ * A token carrying the fields needed to render a link.
+ *
+ * @internal
+ */
+final class AddressToken
+{
+    public function __construct(
+        public readonly bool $isEmail,
+        public readonly string $url,
+        public readonly string $completeUrl,
+        public readonly string $linkText,
+    ) {}
+}